docs update + relational db modules rearranged
