+++ docs/_modules/beam_nuggets/transforms/json_.html
@@ -80,7 +80,7 @@
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="../../../get_started.html">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="../../../get_started.html#usage">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="../../../modules.html">beam_nuggets</a></li>
+<li class="toctree-l1"><a class="reference internal" href="../../../beam_nuggets.html">beam_nuggets package</a></li>
 </ul>
 
             
@@ -151,13 +151,13 @@
 <span class="kn">import</span> <span class="nn">apache_beam</span> <span class="k">as</span> <span class="nn">beam</span>
 
 
-<div class="viewcode-block" id="ParseJson"><a class="viewcode-back" href="../../../beam_nuggets.transforms.html#beam_nuggets.transforms.json_.ParseJson">[docs]</a><span class="k">class</span> <span class="nc">ParseJson</span><span class="p">(</span><span class="n">beam</span><span class="o">.</span><span class="n">DoFn</span><span class="p">):</span>
+<div class="viewcode-block" id="ParseJson"><a class="viewcode-back" href="../../../beam_nuggets.transforms.json_.html#beam_nuggets.transforms.json_.ParseJson">[docs]</a><span class="k">class</span> <span class="nc">ParseJson</span><span class="p">(</span><span class="n">beam</span><span class="o">.</span><span class="n">DoFn</span><span class="p">):</span>
 
     <span class="k">def</span> <span class="nf">__init__</span><span class="p">(</span><span class="bp">self</span><span class="p">,</span> <span class="n">only_keys</span><span class="o">=</span><span class="kc">None</span><span class="p">,</span> <span class="o">*</span><span class="n">unused_args</span><span class="p">,</span> <span class="o">**</span><span class="n">unused_kwargs</span><span class="p">):</span>
         <span class="bp">self</span><span class="o">.</span><span class="n">only_keys</span> <span class="o">=</span> <span class="n">only_keys</span>
         <span class="nb">super</span><span class="p">(</span><span class="n">ParseJson</span><span class="p">,</span> <span class="bp">self</span><span class="p">)</span><span class="o">.</span><span class="fm">__init__</span><span class="p">(</span><span class="o">*</span><span class="n">unused_args</span><span class="p">,</span> <span class="o">**</span><span class="n">unused_kwargs</span><span class="p">)</span>
 
-<div class="viewcode-block" id="ParseJson.process"><a class="viewcode-back" href="../../../beam_nuggets.transforms.html#beam_nuggets.transforms.json_.ParseJson.process">[docs]</a>    <span class="k">def</span> <span class="nf">process</span><span class="p">(</span><span class="bp">self</span><span class="p">,</span> <span class="n">element</span><span class="p">):</span>
+<div class="viewcode-block" id="ParseJson.process"><a class="viewcode-back" href="../../../beam_nuggets.transforms.json_.html#beam_nuggets.transforms.json_.ParseJson.process">[docs]</a>    <span class="k">def</span> <span class="nf">process</span><span class="p">(</span><span class="bp">self</span><span class="p">,</span> <span class="n">element</span><span class="p">):</span>
         <span class="sd">&quot;&quot;&quot;</span>
 <span class="sd">        Args:</span>
 <span class="sd">            element(dict):</span>
