+++ docs/_modules/index.html
@@ -80,7 +80,7 @@
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="../get_started.html">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="../get_started.html#usage">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="../modules.html">beam_nuggets</a></li>
+<li class="toctree-l1"><a class="reference internal" href="../beam_nuggets.html">beam_nuggets package</a></li>
 </ul>
 
             
@@ -143,9 +143,8 @@
             
   <h1>All modules for which code is available</h1>
 <ul><li><a href="beam_nuggets/io/csv_.html">beam_nuggets.io.csv_</a></li>
-<li><a href="beam_nuggets/io/relational_db/read_transform.html">beam_nuggets.io.relational_db.read_transform</a></li>
-<li><a href="beam_nuggets/io/relational_db/sqlalchemy_db.html">beam_nuggets.io.relational_db.sqlalchemy_db</a></li>
-<li><a href="beam_nuggets/io/relational_db/write_transform.html">beam_nuggets.io.relational_db.write_transform</a></li>
+<li><a href="beam_nuggets/io/relational_db.html">beam_nuggets.io.relational_db</a></li>
+<li><a href="beam_nuggets/io/relational_db_api.html">beam_nuggets.io.relational_db_api</a></li>
 <li><a href="beam_nuggets/transforms/json_.html">beam_nuggets.transforms.json_</a></li>
 <li><a href="beam_nuggets/transforms/nested_dict.html">beam_nuggets.transforms.nested_dict</a></li>
 <li><a href="beam_nuggets/transforms/unique_id.html">beam_nuggets.transforms.unique_id</a></li>
