+++ docs/_sources/beam_nuggets.io.relational_db.rst.txt
@@ -1,36 +1,5 @@
-beam\_nuggets.io.relational\_db package
-=======================================
-
-Submodules
-----------
-
-beam\_nuggets.io.relational\_db.read\_transform module
-------------------------------------------------------
-
-.. automodule:: beam_nuggets.io.relational_db.read_transform
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-beam\_nuggets.io.relational\_db.sqlalchemy\_db module
------------------------------------------------------
-
-.. automodule:: beam_nuggets.io.relational_db.sqlalchemy_db
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-beam\_nuggets.io.relational\_db.write\_transform module
--------------------------------------------------------
-
-.. automodule:: beam_nuggets.io.relational_db.write_transform
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
+beam\_nuggets.io.relational\_db module
+======================================
 
 .. automodule:: beam_nuggets.io.relational_db
     :members:
