+++ docs/_sources/beam_nuggets.io.relational_db_api.rst.txt
@@ -1,0 +1,7 @@
+beam\_nuggets.io.relational\_db\_api module
+===========================================
+
+.. automodule:: beam_nuggets.io.relational_db_api
+    :members:
+    :undoc-members:
+    :show-inheritance:
