+++ docs/_sources/beam_nuggets.transforms.nested_dict.rst.txt
@@ -1,0 +1,7 @@
+beam\_nuggets.transforms.nested\_dict module
+============================================
+
+.. automodule:: beam_nuggets.transforms.nested_dict
+    :members:
+    :undoc-members:
+    :show-inheritance:
