+++ docs/_sources/beam_nuggets.transforms.rst.txt
@@ -4,30 +4,11 @@ beam\_nuggets.transforms package
 Submodules
 ----------
 
-beam\_nuggets.transforms.json\_ module
---------------------------------------
+.. toctree::
 
-.. automodule:: beam_nuggets.transforms.json_
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-beam\_nuggets.transforms.nested\_dict module
---------------------------------------------
-
-.. automodule:: beam_nuggets.transforms.nested_dict
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-beam\_nuggets.transforms.unique\_id module
-------------------------------------------
-
-.. automodule:: beam_nuggets.transforms.unique_id
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
+   beam_nuggets.transforms.json_
+   beam_nuggets.transforms.nested_dict
+   beam_nuggets.transforms.unique_id
 
 Module contents
 ---------------
