+++ docs/beam_nuggets.io.html
@@ -28,7 +28,7 @@
   <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
-    <link rel="next" title="beam_nuggets.io.relational_db package" href="beam_nuggets.io.relational_db.html" />
+    <link rel="next" title="beam_nuggets.io.csv_ module" href="beam_nuggets.io.csv_.html" />
     <link rel="prev" title="beam_nuggets package" href="beam_nuggets.html" /> 
 
   
@@ -82,16 +82,17 @@
 <ul class="current">
 <li class="toctree-l1"><a class="reference internal" href="get_started.html">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html#usage">Usage</a></li>
-<li class="toctree-l1 current"><a class="reference internal" href="modules.html">beam_nuggets</a><ul class="current">
-<li class="toctree-l2 current"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a><ul class="current">
-<li class="toctree-l3 current"><a class="reference internal" href="beam_nuggets.html#subpackages">Subpackages</a><ul class="current">
-<li class="toctree-l4 current"><a class="current reference internal" href="#">beam_nuggets.io package</a></li>
-<li class="toctree-l4"><a class="reference internal" href="beam_nuggets.transforms.html">beam_nuggets.transforms package</a></li>
+<li class="toctree-l1 current"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a><ul class="current">
+<li class="toctree-l2 current"><a class="reference internal" href="beam_nuggets.html#subpackages">Subpackages</a><ul class="current">
+<li class="toctree-l3 current"><a class="current reference internal" href="#">beam_nuggets.io package</a><ul>
+<li class="toctree-l4"><a class="reference internal" href="#submodules">Submodules</a></li>
+<li class="toctree-l4"><a class="reference internal" href="#module-beam_nuggets.io">Module contents</a></li>
 </ul>
 </li>
-<li class="toctree-l3"><a class="reference internal" href="beam_nuggets.html#module-beam_nuggets">Module contents</a></li>
+<li class="toctree-l3"><a class="reference internal" href="beam_nuggets.transforms.html">beam_nuggets.transforms package</a></li>
 </ul>
 </li>
+<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.html#module-beam_nuggets">Module contents</a></li>
 </ul>
 </li>
 </ul>
@@ -139,8 +140,6 @@
     
       <li><a href="index.html">Docs</a> &raquo;</li>
         
-          <li><a href="modules.html">beam_nuggets</a> &raquo;</li>
-        
           <li><a href="beam_nuggets.html">beam_nuggets package</a> &raquo;</li>
         
       <li>beam_nuggets.io package</li>
@@ -164,37 +163,15 @@
             
   <div class="section" id="beam-nuggets-io-package">
 <h1>beam_nuggets.io package<a class="headerlink" href="#beam-nuggets-io-package" title="Permalink to this headline">¶</a></h1>
-<div class="section" id="subpackages">
-<h2>Subpackages<a class="headerlink" href="#subpackages" title="Permalink to this headline">¶</a></h2>
-<div class="toctree-wrapper compound">
-<ul>
-<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.io.relational_db.html">beam_nuggets.io.relational_db package</a><ul>
-<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.io.relational_db.html#submodules">Submodules</a></li>
-<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.io.relational_db.html#module-beam_nuggets.io.relational_db.read_transform">beam_nuggets.io.relational_db.read_transform module</a></li>
-<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.io.relational_db.html#module-beam_nuggets.io.relational_db.sqlalchemy_db">beam_nuggets.io.relational_db.sqlalchemy_db module</a></li>
-<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.io.relational_db.html#module-beam_nuggets.io.relational_db.write_transform">beam_nuggets.io.relational_db.write_transform module</a></li>
-<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.io.relational_db.html#module-beam_nuggets.io.relational_db">Module contents</a></li>
-</ul>
-</li>
-</ul>
-</div>
-</div>
 <div class="section" id="submodules">
 <h2>Submodules<a class="headerlink" href="#submodules" title="Permalink to this headline">¶</a></h2>
+<div class="toctree-wrapper compound">
+<ul>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.io.csv_.html">beam_nuggets.io.csv_ module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.io.relational_db.html">beam_nuggets.io.relational_db module</a></li>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.io.relational_db_api.html">beam_nuggets.io.relational_db_api module</a></li>
+</ul>
 </div>
-<div class="section" id="module-beam_nuggets.io.csv_">
-<span id="beam-nuggets-io-csv-module"></span><h2>beam_nuggets.io.csv_ module<a class="headerlink" href="#module-beam_nuggets.io.csv_" title="Permalink to this headline">¶</a></h2>
-<dl class="class">
-<dt id="beam_nuggets.io.csv_.ReadFromCsv">
-<em class="property">class </em><code class="descclassname">beam_nuggets.io.csv_.</code><code class="descname">ReadFromCsv</code><span class="sig-paren">(</span><em>file_name</em>, <em>**kwargs</em><span class="sig-paren">)</span><a class="reference internal" href="_modules/beam_nuggets/io/csv_.html#ReadFromCsv"><span class="viewcode-link">[source]</span></a><a class="headerlink" href="#beam_nuggets.io.csv_.ReadFromCsv" title="Permalink to this definition">¶</a></dt>
-<dd><p>Bases: <code class="xref py py-class docutils literal notranslate"><span class="pre">apache_beam.transforms.ptransform.PTransform</span></code></p>
-<dl class="method">
-<dt id="beam_nuggets.io.csv_.ReadFromCsv.expand">
-<code class="descname">expand</code><span class="sig-paren">(</span><em>pcoll</em><span class="sig-paren">)</span><a class="reference internal" href="_modules/beam_nuggets/io/csv_.html#ReadFromCsv.expand"><span class="viewcode-link">[source]</span></a><a class="headerlink" href="#beam_nuggets.io.csv_.ReadFromCsv.expand" title="Permalink to this definition">¶</a></dt>
-<dd></dd></dl>
-
-</dd></dl>
-
 </div>
 <div class="section" id="module-beam_nuggets.io">
 <span id="module-contents"></span><h2>Module contents<a class="headerlink" href="#module-beam_nuggets.io" title="Permalink to this headline">¶</a></h2>
@@ -209,7 +186,7 @@
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="beam_nuggets.io.relational_db.html" class="btn btn-neutral float-right" title="beam_nuggets.io.relational_db package" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
+        <a href="beam_nuggets.io.csv_.html" class="btn btn-neutral float-right" title="beam_nuggets.io.csv_ module" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
       
       
         <a href="beam_nuggets.html" class="btn btn-neutral" title="beam_nuggets package" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
