+++ docs/get_started.html
@@ -28,7 +28,7 @@
   <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
     <link rel="index" title="Index" href="genindex.html" />
     <link rel="search" title="Search" href="search.html" />
-    <link rel="next" title="beam_nuggets" href="modules.html" />
+    <link rel="next" title="beam_nuggets package" href="beam_nuggets.html" />
     <link rel="prev" title="Welcome to beam-nuggets’ documentation!" href="index.html" /> 
 
   
@@ -82,7 +82,7 @@
 <ul class="current">
 <li class="toctree-l1 current"><a class="current reference internal" href="#">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="#usage">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="modules.html">beam_nuggets</a></li>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a></li>
 </ul>
 
             
@@ -164,7 +164,7 @@
   
     <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
       
-        <a href="modules.html" class="btn btn-neutral float-right" title="beam_nuggets" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
+        <a href="beam_nuggets.html" class="btn btn-neutral float-right" title="beam_nuggets package" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
       
       
         <a href="index.html" class="btn btn-neutral" title="Welcome to beam-nuggets’ documentation!" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
