+++ docs/index.html
@@ -81,7 +81,7 @@
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html#usage">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="modules.html">beam_nuggets</a></li>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a></li>
 </ul>
 
             
@@ -153,8 +153,9 @@
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html#usage">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="modules.html">beam_nuggets</a><ul>
-<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a></li>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a><ul>
+<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.html#subpackages">Subpackages</a></li>
+<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.html#module-beam_nuggets">Module contents</a></li>
 </ul>
 </li>
 </ul>
