+++ docs/modules.html
@@ -27,9 +27,7 @@
   <link rel="stylesheet" href="_static/css/theme.css" type="text/css" />
   <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
     <link rel="index" title="Index" href="genindex.html" />
-    <link rel="search" title="Search" href="search.html" />
-    <link rel="next" title="beam_nuggets package" href="beam_nuggets.html" />
-    <link rel="prev" title="Installation" href="get_started.html" /> 
+    <link rel="search" title="Search" href="search.html" /> 
 
   
   <script src="_static/js/modernizr.min.js"></script>
@@ -79,13 +77,10 @@
             
             
               <p class="caption"><span class="caption-text">Contents:</span></p>
-<ul class="current">
+<ul>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html#usage">Usage</a></li>
-<li class="toctree-l1 current"><a class="current reference internal" href="#">beam_nuggets</a><ul>
-<li class="toctree-l2"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a></li>
-</ul>
-</li>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a></li>
 </ul>
 
             
@@ -157,17 +152,12 @@
 <li class="toctree-l1"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a><ul>
 <li class="toctree-l2"><a class="reference internal" href="beam_nuggets.html#subpackages">Subpackages</a><ul>
 <li class="toctree-l3"><a class="reference internal" href="beam_nuggets.io.html">beam_nuggets.io package</a><ul>
-<li class="toctree-l4"><a class="reference internal" href="beam_nuggets.io.html#subpackages">Subpackages</a></li>
 <li class="toctree-l4"><a class="reference internal" href="beam_nuggets.io.html#submodules">Submodules</a></li>
-<li class="toctree-l4"><a class="reference internal" href="beam_nuggets.io.html#module-beam_nuggets.io.csv_">beam_nuggets.io.csv_ module</a></li>
 <li class="toctree-l4"><a class="reference internal" href="beam_nuggets.io.html#module-beam_nuggets.io">Module contents</a></li>
 </ul>
 </li>
 <li class="toctree-l3"><a class="reference internal" href="beam_nuggets.transforms.html">beam_nuggets.transforms package</a><ul>
 <li class="toctree-l4"><a class="reference internal" href="beam_nuggets.transforms.html#submodules">Submodules</a></li>
-<li class="toctree-l4"><a class="reference internal" href="beam_nuggets.transforms.html#module-beam_nuggets.transforms.json_">beam_nuggets.transforms.json_ module</a></li>
-<li class="toctree-l4"><a class="reference internal" href="beam_nuggets.transforms.html#module-beam_nuggets.transforms.nested_dict">beam_nuggets.transforms.nested_dict module</a></li>
-<li class="toctree-l4"><a class="reference internal" href="beam_nuggets.transforms.html#module-beam_nuggets.transforms.unique_id">beam_nuggets.transforms.unique_id module</a></li>
 <li class="toctree-l4"><a class="reference internal" href="beam_nuggets.transforms.html#module-beam_nuggets.transforms">Module contents</a></li>
 </ul>
 </li>
@@ -185,15 +175,6 @@
            
           </div>
           <footer>
-  
-    <div class="rst-footer-buttons" role="navigation" aria-label="footer navigation">
-      
-        <a href="beam_nuggets.html" class="btn btn-neutral float-right" title="beam_nuggets package" accesskey="n" rel="next">Next <span class="fa fa-arrow-circle-right"></span></a>
-      
-      
-        <a href="get_started.html" class="btn btn-neutral" title="Installation" accesskey="p" rel="prev"><span class="fa fa-arrow-circle-left"></span> Previous</a>
-      
-    </div>
   
 
   <hr/>
