+++ docs/py-modindex.html
@@ -83,7 +83,7 @@
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html#usage">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="modules.html">beam_nuggets</a></li>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a></li>
 </ul>
 
             
@@ -169,7 +169,7 @@
      <tr class="cg-1">
        <td></td>
        <td>&#160;&#160;&#160;
-       <a href="beam_nuggets.io.html#module-beam_nuggets.io.csv_"><code class="xref">beam_nuggets.io.csv_</code></a></td><td>
+       <a href="beam_nuggets.io.csv_.html#module-beam_nuggets.io.csv_"><code class="xref">beam_nuggets.io.csv_</code></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
@@ -179,17 +179,7 @@
      <tr class="cg-1">
        <td></td>
        <td>&#160;&#160;&#160;
-       <a href="beam_nuggets.io.relational_db.html#module-beam_nuggets.io.relational_db.read_transform"><code class="xref">beam_nuggets.io.relational_db.read_transform</code></a></td><td>
-       <em></em></td></tr>
-     <tr class="cg-1">
-       <td></td>
-       <td>&#160;&#160;&#160;
-       <a href="beam_nuggets.io.relational_db.html#module-beam_nuggets.io.relational_db.sqlalchemy_db"><code class="xref">beam_nuggets.io.relational_db.sqlalchemy_db</code></a></td><td>
-       <em></em></td></tr>
-     <tr class="cg-1">
-       <td></td>
-       <td>&#160;&#160;&#160;
-       <a href="beam_nuggets.io.relational_db.html#module-beam_nuggets.io.relational_db.write_transform"><code class="xref">beam_nuggets.io.relational_db.write_transform</code></a></td><td>
+       <a href="beam_nuggets.io.relational_db_api.html#module-beam_nuggets.io.relational_db_api"><code class="xref">beam_nuggets.io.relational_db_api</code></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
@@ -199,17 +189,17 @@
      <tr class="cg-1">
        <td></td>
        <td>&#160;&#160;&#160;
-       <a href="beam_nuggets.transforms.html#module-beam_nuggets.transforms.json_"><code class="xref">beam_nuggets.transforms.json_</code></a></td><td>
+       <a href="beam_nuggets.transforms.json_.html#module-beam_nuggets.transforms.json_"><code class="xref">beam_nuggets.transforms.json_</code></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&#160;&#160;&#160;
-       <a href="beam_nuggets.transforms.html#module-beam_nuggets.transforms.nested_dict"><code class="xref">beam_nuggets.transforms.nested_dict</code></a></td><td>
+       <a href="beam_nuggets.transforms.nested_dict.html#module-beam_nuggets.transforms.nested_dict"><code class="xref">beam_nuggets.transforms.nested_dict</code></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&#160;&#160;&#160;
-       <a href="beam_nuggets.transforms.html#module-beam_nuggets.transforms.unique_id"><code class="xref">beam_nuggets.transforms.unique_id</code></a></td><td>
+       <a href="beam_nuggets.transforms.unique_id.html#module-beam_nuggets.transforms.unique_id"><code class="xref">beam_nuggets.transforms.unique_id</code></a></td><td>
        <em></em></td></tr>
    </table>
