+++ docs/search.html
@@ -80,7 +80,7 @@
 <ul>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html">Installation</a></li>
 <li class="toctree-l1"><a class="reference internal" href="get_started.html#usage">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="modules.html">beam_nuggets</a></li>
+<li class="toctree-l1"><a class="reference internal" href="beam_nuggets.html">beam_nuggets package</a></li>
 </ul>
