+++ scripts/generate_docs.sh
@@ -3,7 +3,7 @@ cd $(dirname "$0")  # going to the same dir as the script
 cd ../sphinx
 
 printf "Generating code documentation .rst files"
-sphinx-apidoc -f -o . ../beam_nuggets/ ../**/test
+sphinx-apidoc -d 4 -e -f -o . ../beam_nuggets/. ../**/test
 
 printf "\nBuilding html"
 make clean html
