+++ sphinx/beam_nuggets.io.csv_.rst
@@ -1,0 +1,7 @@
+beam\_nuggets.io.csv\_ module
+=============================
+
+.. automodule:: beam_nuggets.io.csv_
+    :members:
+    :undoc-members:
+    :show-inheritance:
