+++ sphinx/beam_nuggets.transforms.unique_id.rst
@@ -1,0 +1,7 @@
+beam\_nuggets.transforms.unique\_id module
+==========================================
+
+.. automodule:: beam_nuggets.transforms.unique_id
+    :members:
+    :undoc-members:
+    :show-inheritance:
